[SNOW-1156046] fix toctou vulnerability in EasyLogginConfig
